requirements.txt
torch
torchvision
opencv-python
numpy
matplotlib
tqdm
tensorflow
scikit-learn
streamlit
opencv-python-headless
pillow